tarea2/prestamos/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.3
cross-web==0.6.0
Flask==3.1.3
graphql-core==3.2.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.2
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
six==1.17.0
strawberry-graphql==0.315.0
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
